pyproject.toml
[project]
name = "bigquery_agents"
version = "0.1.13"
description = ""
authors = [
    {name = "sametkaradag",email = "sametkaradag@gmail.com"}
]
readme = "README.md"
license = "Google"
requires-python = ">=3.12,<4.0"
dependencies = [
    "google-cloud-bigquery (>=3.29.0,<4.0.0)",
    "langchain (>=0.3.17,<0.4.0)",
    "langgraph (>=0.2.69,<0.3.0)",
    "langchain-google-vertexai (>=2.0.12,<3.0.0)",
    "ipython (>=8.31.0,<9.0.0)",
]
packages = [
    {include = "bq_fix", from = "src/bigquery_agents"},
    {include = "bq_fix_basic", from = "src/bigquery_agents"},
    {include = "bq_medic", from = "src/bigquery_agents"}
]

[tool.poetry.scripts]
bq_fix_basic = "bigquery_agents.bq_fix_basic.__main__:main"
bq_fix = "bigquery_agents.bq_fix.__main__:main"
bq_medic = "bigquery_agents.bq_medic:main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
